Move `PlainTextFormatter` to `cli_fmt` and `Train` to `rt_logic`.
