+++ crate/cli_fmt/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "choochoo_cli_fmt"
+version = "0.1.0"
+authors = ["Azriel Hoh <azriel91@gmail.com>"]
+edition = "2018"
+description = "Command line interface formatting logic for the choochoo automation library."
+repository = "https://github.com/azriel91/choochoo"
+documentation = "https://docs.rs/choochoo_rt_logic/"
+readme = "../../README.md"
+keywords = ["automation"]
+license = "MIT OR Apache-2.0"
+
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+choochoo_cfg_model = { path = "../cfg_model", version = "0.1.0" }
+choochoo_rt_model = { path = "../rt_model", version = "0.1.0" }
+futures = "0.3.17"
+tokio = { version = "1.11.0", features = ["io-util", "rt", "sync"] }
