+++ crate/cli_fmt/src/lib.rs
@@ -1,0 +1,5 @@
+//! Command line interface formatting logic for the choochoo automation library.
+
+pub use crate::plain_text_formatter::PlainTextFormatter;
+
+mod plain_text_formatter;
