+++ crate/rt_logic/src/lib.rs
@@ -1,8 +1,9 @@
 //! Runtime visit logic for the choochoo automation library.
 
-pub use crate::{driver::Driver, visit_status_updater::VisitStatusUpdater};
+pub use crate::{driver::Driver, train::Train, visit_status_updater::VisitStatusUpdater};
 
 pub mod strategy;
 
 mod driver;
+mod train;
 mod visit_status_updater;
