--- src/fmt.rs
@@ -1,5 +1,0 @@
-//! Formats to display information to the user.
-
-pub use self::plain_text_formatter::PlainTextFormatter;
-
-mod plain_text_formatter;
