+++ workspace_tests/Cargo.toml
@@ -17,6 +17,7 @@ test = true
 
 [dependencies]
 choochoo_cfg_model = { path = "../crate/cfg_model", version = "0.1.0" }
+choochoo_cli_fmt = { path = "../crate/cli_fmt", version = "0.1.0" }
 choochoo_rt_logic = { path = "../crate/rt_logic", version = "0.1.0" }
 choochoo_rt_model = { path = "../crate/rt_model", version = "0.1.0" }
 tokio = { version = "1.11.0", features = ["io-util", "rt", "sync"] }
