+++ workspace_tests/src/cli_fmt.rs
@@ -1,0 +1,1 @@
+mod plain_text_formatter;
