+++ workspace_tests/src/lib.rs
@@ -3,5 +3,6 @@
 //! Tests for the choochoo automation library.
 
 mod cfg_model;
+mod cli_fmt;
 mod rt_logic;
 mod rt_model;
