+++ workspace_tests/src/rt_logic.rs
@@ -1,2 +1,3 @@
 mod integrity_strat;
+mod train;
 mod visit_status_updater;
